requirements.txt
anthropic>=0.40.0
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.0
python-dateutil>=2.8.2
